[kubevirt] Move harvester files from models directory
harvester directory has been moved from models to top level,
this fixes package build and makes Rancher shell models auto import
work correctly
