+++ .prettierignore
@@ -1,2 +1,2 @@
-pkg/kubevirt-dashboard-extension/models/harvester/**
+pkg/kubevirt-dashboard-extension/harvester/**
