+++ pkg/kubevirt-dashboard-extension/harvester/config/harvester.ts
