+++ pkg/kubevirt-dashboard-extension/harvester/config/labels-annotations.js
