+++ pkg/kubevirt-dashboard-extension/harvester/config/types.js
