+++ pkg/kubevirt-dashboard-extension/harvester/types.ts
