+++ pkg/kubevirt-dashboard-extension/harvester/utils/vm.js
